Step 2: add main app

+++ main-service/src/main/java/ru/practicum/categories/service/DefaultCategoryService.java
@@ -10,11 +10,14 @@ import ru.practicum.categories.dto.CategoryDtoOut;
 import ru.practicum.categories.mapper.CategoryMapper;
 import ru.practicum.categories.model.Category;
 import ru.practicum.categories.repository.CategoryRepository;
+import ru.practicum.event.model.Event;
+import ru.practicum.event.repository.EventsRepository;
 import ru.practicum.exception.model.BadRequestException;
 import ru.practicum.exception.model.ConflictException;
 import ru.practicum.exception.model.NotFoundException;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 
@@ -23,6 +26,7 @@ import java.util.stream.Collectors;
 public class DefaultCategoryService implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventsRepository eventsRepository;
 
     @Override
     public CategoryDtoOut saveCategory(CategoryDtoIn categoryDtoIn) {
@@ -34,11 +38,11 @@ public class DefaultCategoryService implements CategoryService {
 
     @Override
     public CategoryDtoOut updateCategory(CategoryDtoIn categoryDtoIn, long id) {
-        Category oldCategory = categoryRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException(String.format("Категория с id = %d не найдена", id)));
-        if (oldCategory.getName().equals(categoryDtoIn.getName())) {
+        if (checkNameEvent(categoryDtoIn, id)) {
             throw new ConflictException("Категория с таким именем уже существует");
         }
+        Category oldCategory = categoryRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException(String.format("Категория с id = %d не найдена", id)));
         oldCategory.setName(categoryDtoIn.getName());
         return CategoryMapper.toCategoryOut(categoryRepository.save(oldCategory));
     }
@@ -47,6 +51,13 @@ public class DefaultCategoryService implements CategoryService {
     public void deleteCategory(long id) {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException(String.format("Категория с id = %d не найдена", id)));
+        boolean check = eventsRepository.findAll().stream()
+                .map(Event::getCategory)
+                .map(Category::getId)
+                .anyMatch(ids -> id == ids);
+        if (check) {
+            throw new ConflictException("Нельзя удалять категорию");
+        }
         categoryRepository.delete(category);
 
     }
@@ -76,5 +87,11 @@ public class DefaultCategoryService implements CategoryService {
         return categoryRepository.findAll().stream()
                 .map(Category::getName)
                 .anyMatch(name -> name.equals(categoryDtoIn.getName()));
+    }
+
+    private boolean checkNameEvent(CategoryDtoIn categoryDtoIn, Long ids) {
+        return categoryRepository.findAll().stream()
+                .anyMatch(category -> category.getName().equals(categoryDtoIn.getName())
+                        && !Objects.equals(category.getId(), ids));
     }
 }
